Update vidogl.c
New trial to fix the malformed VDP1 coordinates. It looks like not all the MSB bits are analysed by the real vdp1 but only some of them. this is an issue in standard 16bit and 32 bits C style representation. te patch is first fixing the bit sign extension (not sure it is enough) and then it is clamping the coordiantes (not sure it is needed).